digital_assets/Swarm_Sales_Infiltrator/agent.py
"""
DIGITAL SWARM | ELITE ASSET COLLECTION
ID: Swarm Sales Infiltrator - Core Agent
VERSION: 1.0.0
"""

import os
from dotenv import load_dotenv
from tools import generate_battle_card_html

load_dotenv()


class SalesInfiltratorAgent:
    def __init__(self, agent_id="SWARM-UNIT-07"):
        self.id = agent_id
        print(f"🛰️ {self.id} UPLINK ESTABLISHED.")

    def run_infiltration(self, product_name, competitor_name):
        print(f"🔍 INFILTRATING COMPETITOR DATA: {competitor_name}...")

        # Phase 1-7 (Simulated multi-agent logic for this production version)
        intel = f"""
        COMPETITIVE POSITIONING:
        - Product: {product_name}
        - Target: {competitor_name}
        
        SWOT ANALYSIS:
        - STRENGTHS: Modular architecture, Cyberpunk UI.
        - WEAKNESSES: Emerging brand presence.
        - OPPORTUNITIES: Deep RAG integration.
        - THREATS: Established legacy players.
        
        OBJECTION SCRIPTS:
        - "It's too technical" -> Response: "Digital Swarm assets are designed for rapid deployment, abstracting away the complexity for elite performance."
        """

        print("🧪 SYNTHESIZING BATTLE CARD...")
        filename = generate_battle_card_html(intel, product_name, competitor_name)

        print(f"✅ INFILTRATION COMPLETE. ASSET DEPLOYED: {filename}")
        return filename


if __name__ == "__main__":
    agent = SalesInfiltratorAgent()
    p_name = input("ENTER YOUR PRODUCT NAME: ")
    c_name = input("ENTER COMPETITOR NAME: ")
    agent.run_infiltration(p_name, c_name)
