digital_assets/Swarm_Sales_Infiltrator/requirements.txt
python-dotenv
streamlit
# Note: This version assumes standard HTTP requests/LLM API calls for compatibility
requests
openai
